@@ -2,27 +2,13 @@
   "name": "rick-n-morty",
   "version": "0.1.0",
   "private": true,
-  "dependencies": {
-    "@popperjs/core": "^2.11.0",
-    "@testing-library/jest-dom": "^5.11.4",
-    "@testing-library/react": "^11.1.0",
-    "@testing-library/user-event": "^12.1.10",
-    "@types/jest": "^29.5.14",
-    "@types/react": "^18.3.12",
-    "@types/react-dom": "^18.3.1",
-    "bootstrap": "^5.1.3",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-paginate": "^8.0.2",
-    "react-router-dom": "^6.0.2",
-    "react-scripts": "4.0.3",
-    "sass": "^1.44.0"
-  },
   "scripts": {
-    "start": "react-scripts start",
+    "start": "cross-env NODE_OPTIONS=--openssl-legacy-provider react-scripts start",
     "build": "react-scripts build",
     "test": "react-scripts test",
-    "eject": "react-scripts eject"
+    "eject": "react-scripts eject",
+    "tsc": "tsc src",
+    "format": "prettier --write ."
   },
   "eslintConfig": {
     "extends": [
@@ -42,7 +28,27 @@
       "last 1 safari version"
     ]
   },
+  "dependencies": {
+    "@popperjs/core": "^2.11.0",
+    "bootstrap": "^5.1.3",
+    "cross-env": "^7.0.3",
+    "react": "18.2.0",
+    "react-dom": "18.2.0",
+    "react-paginate": "^8.2.0",
+    "react-router-dom": "6.22.0",
+    "react-scripts": "^5.0.1",
+    "sass": "^1.44.0"
+  },
   "devDependencies": {
+    "@testing-library/jest-dom": "^6.6.3",
+    "@testing-library/react": "^16.0.1",
+    "@testing-library/user-event": "^14.5.2",
+    "@types/jest": "^29.5.14",
+    "@types/react": "^18.3.12",
+    "@types/react-dom": "^18.3.1",
+    "jest": "^27.5.1",
+    "prettier": "^3.3.3",
+    "testing-library": "^0.0.2",
     "typescript": "^5.6.3"
   }
 }
