@@ -1,11 +1,12 @@
-import React from "react"
-import ReactDOM from "react-dom"
-import "./index.css"
-import App from "./App"
+import React from "react";
+import { createRoot } from "react-dom/client";
+import { App } from "./app/App";
+import "./index.css";
 
-ReactDOM.render(
+const container = document.getElementById("app");
+const root = createRoot(container);
+root.render(
   <React.StrictMode>
     <App />
-  </React.StrictMode>,
-  document.getElementById("root")
-)
+  </React.StrictMode>
+);
